Update _JSON_Formator.py
Updated the script to work with defaulted values of CREATOR and ID.

Added restriction of having only one blank space in a text question.

fixed IndexError : of when a user directory exists but has no files within.

Added exception to after the first iteration of adding question the code will try to pop a key such as options. which wil not be there. hence KeyError :